Add methods for computing the dot/dash soundbite duration in seconds.
Assuming you know the delay of a single dot sound to be X, then when
translating to Morse code you need to insert the following delays in the
output: X delay between the chars in the Morse encoding, 3 * X delay
between the char encodings, and 7 * X between words. A method that tells
us the duration of dot is then necessary to implement these delays later
on.